backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
twilio
google-genai
python-multipart
